[refactore/#177] 오늘의 루틴페이지에서 습관 체크, 달성률에 따른 상태 변화 유스케이스 하나로 통합
